fix: Id's properly increment with submitted data
The server was automatically incrementing data, which was unknown before and the frontend was attempting to create the id with a count of fetch objects
This really messed up the increment count and prevented the proper number of reviews from being pulled.
Frontend id handling has been removed and no longer causes any issue.