feat: form data is submitted when back online
If a server goes offline or internet connection is disrupted, the DBHelper.submitReviews() function opens a new idb called 'reviews-to-submit' and writes the review data ad an idb object
When a new page loads, a function is initiated to check if the idb has any objects.
If it does, the ojbects are written to the sails server .db and then deleted from the database.
If there are no objects, the function returns and ends.